// types/team.ts
export interface Team {
  idTeam: string;
  strTeam: string;
  strTeamAlternate?: string;
  strTeamShort?: string;
  intFormedYear?: string;
  strSport?: string;
  strLeague?: string;
  idLeague?: string;
  strStadium?: string;
  strStadiumLocation?: string;
  strStadiumCapacity?: string;
  strWebsite?: string;
  strFacebook?: string;
  strTwitter?: string;
  strInstagram?: string;
  strDescriptionEN?: string;
  strCountry?: string;
  strBadge?: string;
  strJersey?: string;
  strLogo?: string;
  strFanart1?: string;
  strBanner?: string;
}